refactor: hide plugin internals
