+++ plugin/src/main/java/io/lonmstalker/tgkit/plugin/BotPluginManager.java
@@ -1,12 +1,15 @@
 package io.lonmstalker.tgkit.plugin;
 
-import static io.lonmstalker.tgkit.plugin.BotPluginConstants.CURRENT_VERSION;
+import static io.lonmstalker.tgkit.plugin.internal.BotPluginConstants.CURRENT_VERSION;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import com.github.zafarkhaja.semver.Version;
 import io.lonmstalker.tgkit.core.exception.BotApiException;
-import io.lonmstalker.tgkit.plugin.sort.TopoSorter;
+import io.lonmstalker.tgkit.plugin.internal.BotPluginContainer;
+import io.lonmstalker.tgkit.plugin.internal.BotPluginContextDefault;
+import io.lonmstalker.tgkit.plugin.internal.BotPluginDescriptor;
+import io.lonmstalker.tgkit.plugin.internal.sort.TopoSorter;
 import io.lonmstalker.tgkit.security.audit.AuditBus;
 import io.lonmstalker.tgkit.security.audit.AuditEvent;
 import io.lonmstalker.tgkit.security.config.BotSecurityGlobalConfig;
