+++ plugin/src/main/java/io/lonmstalker/tgkit/plugin/internal/BotPluginConstants.java
