+++ plugin/src/main/java/io/lonmstalker/tgkit/plugin/internal/BotPluginDescriptor.java
