+++ plugin/src/main/java/io/lonmstalker/tgkit/plugin/internal/ChildFirstURLClassLoader.java
