+++ plugin/src/main/java/io/lonmstalker/tgkit/plugin/internal/Internal.java
@@ -1,0 +1,13 @@
+package io.lonmstalker.tgkit.plugin.internal;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/** Marks API elements as internal and subject to change without notice. */
+@Documented
+@Retention(RetentionPolicy.CLASS)
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface Internal {}
