+++ plugin/src/test/java/io/lonmstalker/tgkit/plugin/BotPluginManagerTest.java
@@ -1,6 +1,6 @@
 package io.lonmstalker.tgkit.plugin;
 
-import static io.lonmstalker.tgkit.plugin.BotPluginConstants.CURRENT_VERSION;
+import static io.lonmstalker.tgkit.plugin.internal.BotPluginConstants.CURRENT_VERSION;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.Mockito.*;
@@ -8,6 +8,7 @@ import static org.mockito.Mockito.*;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import io.lonmstalker.tgkit.core.init.BotCoreInitializer;
+import io.lonmstalker.tgkit.plugin.internal.BotPluginDescriptor;
 import io.lonmstalker.tgkit.security.audit.AuditBus;
 import io.lonmstalker.tgkit.security.audit.AuditEvent;
 import io.lonmstalker.tgkit.security.config.BotSecurityGlobalConfig;
